refactor(frontend/cart): add cart module
