+++ packages/frontend/src/components/containers/index.ts
@@ -1,5 +1,5 @@
 import PayPal from '../modules/Payment/PayPal';
-import CartContent from './CartContent';
+import CartContent from '../modules/Cart/CartContent';
 import UserForms from './UserForms';
 import UserProducts from './UserProducts';
