+++ packages/frontend/src/components/modules/Cart/Cart.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import '../../styles/toastify.css';
+import '../../../styles/toastify.css';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
-import { CartContent } from '../containers/';
-import { Nav } from '../elements/';
+import CartContent from './CartContent';
+import { Nav } from '../../elements';
 
 const Cart = () => {
   return (
