+++ packages/frontend/src/components/modules/Cart/CartCard.tsx
@@ -1,16 +1,16 @@
 import React, { Dispatch, useContext, useState } from 'react';
-import '../../styles/molecules/cartCard.css';
-import { HTTPException } from '../../utils/HttpException';
+import '../../../styles/molecules/cartCard.css';
+import { HTTPException } from '../../../utils/HttpException';
 import RemoveIcon from '../../img/remove.svg';
 import Cookie from 'universal-cookie';
-import { API_KEY } from '../../config';
-import CartContext from '../../context/cart/cart';
-import { removeFromCart } from '../../context/cart/actionsCreator';
-import { IAction, IInitialState } from '../../context/interfaces';
+import { API_KEY } from '../../../config';
+import CartContext from '../../../context/cart/cart';
+import { removeFromCart } from '../../../context/cart/actionsCreator';
+import { IAction, IInitialState } from '../../../context/interfaces';
 import { Link } from 'react-router-dom';
 import { toast } from 'react-toastify';
-import { put } from '../../utils/petitions';
-import { Loader } from '../elements';
+import { put } from '../../../utils/petitions';
+import { Loader } from '../../elements';
 
 const CartCard = ({ img, price, name, _id, frontID }: ICartCard) => {
   const [state, dispatch] = useContext(CartContext) as [IInitialState, Dispatch<IAction>];
