+++ packages/frontend/src/components/molecules/index.ts
@@ -1,6 +1,5 @@
 import Searcher from './Searcher';
-import CartCard from './CartCard';
 import BasicInfo from './BasicInfo';
 import ListProducts from './ListProducts';
 
-export { Searcher, CartCard, BasicInfo, ListProducts };
+export { Searcher, BasicInfo, ListProducts };
