+++ packages/frontend/src/components/pages/index.ts
@@ -1,7 +1,6 @@
 import Main from './Main';
-import Cart from './Cart';
 import User from './User';
 import UserProducts from './UserProducts';
 import UserCart from './UserCart';
 
-export { Main, Cart, User, UserProducts, UserCart };
+export { Main, User, UserProducts, UserCart };
